Extract helpers and fix "allow dead code" in common fonts
