+++ src/common/fonts.rs
@@ -1,4 +1,4 @@
-#[allow(dead_code)]
+#![allow(dead_code)]
 
 pub const SPACE_MONO_REGULAR: &str = "SpaceMono-Regular";
